% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListNearFacebookVenues.R
\name{getListNearFacebookVenues}
\alias{getListNearFacebookVenues}
\title{A getListNearFacebookVenues Function}
\usage{
getListNearFacebookVenues(lat, lon, distance = 1000, q = "")
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{distance}{distance}
}
\description{
This function return geolocated facebook pages
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
