% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListNearVenues.R
\name{getListNearVenues}
\alias{getListNearVenues}
\title{A getListNearVenues Function}
\usage{
getListNearVenues(latVenue, lonVenue, radius = 500)
}
\arguments{
\item{latVenue}{latitude}

\item{lonVenue}{longitude}

\item{radius}{search radius in meteres}
}
\description{
This function return data frame with 50 venues around GPS point
}
\examples{
lat <- "50.0850165"
lon <- "14.4212446"
exampleVenues <- getListNearVenues(lat,lon,radius = 100)

}
