% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfSimilaryVenues.R
\name{getListOfSimilaryVenues}
\alias{getListOfSimilaryVenues}
\title{A getListOfSimilaryVenues Function}
\usage{
getListOfSimilaryVenues(idVenue, log = "TRUE")
}
\arguments{
\item{idVenue}{4SQ id}
}
\description{
This function return data frame with ids of similary venues.
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
