% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListSearchVenues.R
\name{getListSearchVenues}
\alias{getListSearchVenues}
\title{A getListSearchVenues Function}
\usage{
getListSearchVenues(latVenue, lonVenue, search, radius = 5000, limit = 50)
}
\arguments{
\item{latVenue}{latitude}

\item{lonVenue}{longitude}

\item{search}{query}

\item{radius}{distance in meters}

\item{limit}{limit of results}
}
\description{
This function return results for search in 4sq venues names
}
\examples{
latVenue <- "50.08398"
lonVenue <- "14.41663"
kavarny <- getListSearchVenues(latVenue, lonVenue, "coffee", radius = 2000)

}
\keyword{printLog}
