% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVenueDetail.R
\name{getVenueDetail}
\alias{getVenueDetail}
\title{A getVenueDetail Function}
\usage{
getVenueDetail(idVenue, log = "TRUE")
}
\arguments{
\item{idVenue}{4sq venue detail}
}
\description{
This function return 4sq venue details
}
\examples{

getVenueDetail("4bb135eaf964a520c2833ce3")

}
\keyword{printLog}
