% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVenueDetailMultimedia.R
\name{getVenueDetailMultimedia}
\alias{getVenueDetailMultimedia}
\title{A getVenueDetailMultimedia Function}
\usage{
getVenueDetailMultimedia(idVenue)
}
\arguments{
\item{idVenue}{id of 4sq venue}
}
\description{
This function return list of photos from 4sq venue
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
