% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWikipediaCategory.R
\name{getWikipediaCategory}
\alias{getWikipediaCategory}
\title{A getWikipediaLangList Function}
\usage{
getWikipediaCategory(pageName, lang = "cs")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

