% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWikipediaVenuesGPS.R
\name{getWikipediaVenuesGPS}
\alias{getWikipediaVenuesGPS}
\title{A getWikipediaVenuesGPS Function}
\usage{
getWikipediaVenuesGPS(lat, lon, distance, lang = "cs")
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{distance}{in meters}

\item{lang}{language (default is cs)}
}
\description{
This return geolocated wiki pages.
}
\examples{
lat <- "50.02912"
lon <- "13.36412"
zihle <- getWikipediaVenuesGPS(lat,lon,"10000")

}
\keyword{printLog}
