% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipediaCount.R
\name{wikipediaCount}
\alias{wikipediaCount}
\title{A wikipediaCount Function}
\usage{
wikipediaCount(name, lang, from = "2014-01-01", to = "2016-05-20")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
