% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomatrix.R
\name{geomatrix}
\alias{geomatrix}
\title{A getVenueDetail Function}
\usage{
geomatrix(lat, lon, distanceX, distanceY, steps, radius = 90)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
