% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListNearFacebookVenues.R
\name{getListNearFacebookVenues}
\alias{getListNearFacebookVenues}
\title{A getListNearVenues Function}
\usage{
getListNearFacebookVenues(lat, lon, distance = 1000, q = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

