% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListNearVenues.R
\name{getListNearVenues}
\alias{getListNearVenues}
\title{A getListNearVenues Function}
\usage{
getListNearVenues(latVenue, lonVenue, radius = 300)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

