% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfNextVenues.R
\name{getListOfNextVenues}
\alias{getListOfNextVenues}
\title{A getListOfNextVenues Function}
\usage{
getListOfNextVenues(idVenue, log = "TRUE")
}
\arguments{
\item{idVenue}{4sq venue id}
}
\description{
This function return next venues on 4SQ
}
\examples{
getListOfNextVenues("4bb135eaf964a520c2833ce3")


}
\keyword{printLog}
