% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListSearchVenues.R
\name{getListSearchVenues}
\alias{getListSearchVenues}
\title{A getListSearchVenues Function}
\usage{
getListSearchVenues(latVenue, lonVenue, search, radius = 50000, limit = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}
