% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWikipediaVenuesGPS.R
\name{getWikipediaVenuesGPS}
\alias{getWikipediaVenuesGPS}
\title{A getWikipediaVenuesGPS Function}
\usage{
getWikipediaVenuesGPS(lat, lon, distance, lang = "cs")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows pretty printing of values
}
\examples{
a <- "1234"
cat_function(a)

}
\keyword{printLog}

